api/requirements.txt
PyJWT==2.8.0
livekit==0.11.0
python-dotenv==1.0.0 